rewrite remained js files to tsx
