--- frontend/src/javascripts/components/atoms/planet-img.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const PlanetImg = ({ src }) => <img className="planet" src={src} />
-
-export default PlanetImg
