+++ frontend/src/javascripts/components/atoms/planet-img.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react'
+
+interface PlanetImgProps {
+  src: string
+}
+
+const PlanetImg: React.SFC<PlanetImgProps> = ({ src = '' }) => <img className="planet" src={src} />
+
+export default PlanetImg
