+++ frontend/tsconfig.json
@@ -62,6 +62,10 @@
   "exclude": [
     "node_modules",
     "dist",
-    "webpack.config.js" // これしないとエラー起きるけどそこ直せるなら削除した方がいいかも
+    "coverage",
+    "webpack.config.js", // これしないとエラー起きるけどそこ直せるなら削除した方がいいかも
+    "src/jest.config.js",
+    "src/__mocks__",
+    "src/__tests__"
   ]
 }
